Update touchable.js
Changed e.preventDefauling from being used on touchstart to be used on touchmove, this way events like link clicks can continue to be touchable without losing their default behaviors. If people want to prevent some default behavior they can do it externally, not in the plugin. Because is impossible to undo a defaultbehavior once it has been set.

Although an alternative is to simulate a click, in case of links, simulate a click will be kind of difficult (i know you can trigger a click event, but it won't have the security permissions and stuff like simulate a Ctrl+Click or a scroll click will be an innecesary pain).

Also detected that on iOS Mobile Safari browsers both touchstart and mousdown handlers were being invoked, so if an animation was fired on touchstart, after some delay mousedown triggered the animation again. I think it is cleaner to have just one trigger depending on the browser capabilities better than having a dynamic set of callbacks firing inconsistently across browsers.

Additionally: Fixed some code formatting and order (personal mental issues with indenting and ordering).
Additionally: The dynamic usage of eventListeners, tried to have them uniformed all using jQuery bindings, there is no need to recode the cross browser eventListener support if this is actually a jQuery plugin and is running with jQuery. I know this change conflicts with the comment in the summary explaining to users that this plugin can easily be ported to other libraries like Zepto, maybe not with this change, but this can be roll-backed.